changes output format back to uint16 to save memory and storage

+++ eosdxanalysis/preprocessing/preprocess.py
@@ -472,13 +472,13 @@ class PreprocessData(object):
                 if output_format == "txt":
                     save_filename = "{}_{}".format(output_style_abbreviation, filename)
                     save_filename_fullpath = os.path.join(plan_output_dir, save_filename)
-                    np.savetxt(save_filename_fullpath,output.astype(np.uint32),fmt='%i')
+                    np.savetxt(save_filename_fullpath,output.astype(np.uint16),fmt='%i')
 
                 # Save output as image
                 if output_format == "png":
                     save_filename = "{}_{}.png".format(output_style_abbreviation, filename)
                     save_filename_fullpath = os.path.join(plan_output_dir, save_filename)
-                    imageio.imwrite(save_filename_fullpath, output.astype(np.uint32))
+                    imageio.imwrite(save_filename_fullpath, output.astype(np.uint16))
 
 
 if __name__ == "__main__":
